2nd Checkin

+++ src/app/Team/team.component.ts
@@ -1,9 +1,8 @@
 import { Component, OnInit, Injectable } from '@angular/core';
-import { FormBuilder, FormGroup, Validators, FormControl, FormArray  } from '@angular/forms';
+import { FormBuilder, FormGroup, Validators  } from '@angular/forms';
+
 import { TeamService } from '../Services/team.service';
 import { Team } from '../Models/team.model'
-
-
 
 @Component({
   selector: 'app-team',
@@ -12,55 +11,39 @@ import { Team } from '../Models/team.model'
 })
 
 @Injectable()
-export class TeamComponent implements OnInit {
-  //public API = 'https://localhost:44399/api'
-  //teamForm = new Team('', '', '', '', '', '', 5);
-   public teamForm: FormGroup;
-  // dataSaved = false;
-  // massage = null;
-  // employeeIdUpdate = null;
-  teams: Team[];
 
-  constructor(private formBuilder: FormBuilder, private teamService: TeamService)
-  {
+export class TeamComponent implements OnInit{
+  public teamForm: FormGroup;
+  teams: Team[]
+  public routerLinkVariable = "/app-team-detail"; // the value of the variable is string!
+
+  constructor(private formBuilder: FormBuilder, private teamService: TeamService) { }
 
 
-    }
+  ngOnInit() {
+    this.teamForm = this.formBuilder.group({
+      Name: ['', [Validators.required]],
+      State: ['', [Validators.required]],
+      City: ['', [Validators.required]],
+      Owner: ['', [Validators.required]],
+      Coach: ['', [Validators.required]],
+      NumberOfSuperBowlTrophies: ['', [Validators.required]],
+      TeamColors: ['', [Validators.required]]
+    });
 
-    ngOnInit() {
-      this.teamForm = this.formBuilder.group({
-        Name: ['', [Validators.required]],
-        State: ['', [Validators.required]],
-        City: ['', [Validators.required]],
-        Owner: ['', [Validators.required]],
-        Coach: ['', [Validators.required]],
-        NumberOfSuperBowlTrophies: ['', [Validators.required]],
-        TeamColors: ['', [Validators.required]]
-      });
-
-      this.getTeams();
-    }
-
-    getTeams(){
-      this.teamService.getTeams().subscribe(teams => {this.teams = teams; console.log(this.teams)});
-    }
-
-    onSubmit()
-    {
-       const team = this.teamForm.value;
-       this.createTeam(team);
-       this.teams.push(team);
-    }
-
-    createTeam(team: Team){
-      this.teamService.createTeam(team).subscribe(
-          (res) => console.log(res),
-          (err) => console.log(err)
-        );
-   }
-
-   deleteTeam(team: Team): void {
-    this.teams = this.teams.filter(h => h !== team);
-    this.teamService.deleteTeam(team).subscribe();
+    this.getTeams();
   }
+
+  getTeams(){
+    this.teamService.getTeams().subscribe(teams => {this.teams = teams; console.log(this.teams)});
+  }
+
+  deleteTeam(team: Team): void {
+    if (confirm("Are you sure you want to delete team " + team.Name))
+    {
+      this.teams = this.teams.filter(h => h !== team);
+      this.teamService.deleteTeam(team).subscribe();
+    }
+  }
+
 }
